Implemented asynchronous songs loading logic (using mock)
Added status bar and a button for loading songs. The button is used as a UI element to initiate the asynchronous action. It is probably a good idea to remove this button in future.

Rearranged the structure of the project, renamed some components. Moved mock songs to a separate file.

Added 'redux-thunk' middleware.